requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hypothesis==6.145.0
idna==3.11
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pydantic==2.11.5
pydantic-extra-types==2.10.6
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.3
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
starlette==0.46.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.34.0
watchfiles==1.1.1
websockets==15.0.1
